+++ kspcalc.js
@@ -70,21 +70,21 @@ var PACKS = [
 		name:"Stock 1.0",
 		parts:[
 			//LF/O Engines
-			{"name":"S3 KS-25x4 \"Mammoth\" Liquid Fuel Engine","type":TYPES.LFO_ENGINE,"size":3,"cost":39000,"mass":15,"thrust_min":0,"thrust_max":4000,"throttleable":true,"isp_vac":315,"isp_atm":295,"gimbal":2,"last":true},
-			{"name":"Kerbodyne KR-2L+ \"Rhino\" Liquid Fuel Engine","type":TYPES.LFO_ENGINE,"size":3,"cost":21000,"mass":8.5,"thrust_min":0,"thrust_max":2000,"throttleable":true,"isp_vac":340,"isp_atm":170,"gimbal":4},
-			{"name":"RE-M3 \"Mainsail\" Liquid Engine","type":TYPES.LFO_ENGINE,"size":2,"cost":13000,"mass":6,"thrust_min":0,"thrust_max":1500,"throttleable":true,"isp_vac":310,"isp_atm":285,"gimbal":2},
-			{"name":"RE-I5 \"Skipper\" Liquid Fuel Engine","type":TYPES.LFO_ENGINE,"size":2,"cost":5300,"mass":3,"thrust_min":0,"thrust_max":650,"throttleable":true,"isp_vac":320,"isp_atm":280,"gimbal":2},
-			{"name":"RE-L10 \"Poodle\" Liquid Fuel Engine","type":TYPES.LFO_ENGINE,"size":2,"cost":1300,"mass":1.75,"thrust_min":0,"thrust_max":220,"throttleable":true,"isp_vac":350,"isp_atm":85,"gimbal":4.5},
-			{"name":"CR-7 R.A.P.I.E.R. Engine","type":TYPES.LFO_ENGINE,"size":1,"cost":6000,"mass":2,"thrust_min":0,"thrust_max":180,"throttleable":true,"isp_vac":305,"isp_atm":275,"gimbal":3},
-			{"name":"LV-T45 \"Swivel\" Liquid Fuel Engine","type":TYPES.LFO_ENGINE,"size":1,"cost":1200,"mass":1.5,"thrust_min":0,"thrust_max":200,"throttleable":true,"isp_vac":320,"isp_atm":270,"gimbal":3},
-			{"name":"T-1 Toroidal \"Aerospike\" Liquid Fuel Engine","type":TYPES.LFO_ENGINE,"size":1,"cost":3850,"mass":1.5,"thrust_min":0,"thrust_max":175,"throttleable":true,"isp_vac":330,"isp_atm":290,"last":true},
-			{"name":"LV-T30 \"Reliant\" Liquid Fuel Engine","type":TYPES.LFO_ENGINE,"size":1,"cost":1100,"mass":1.25,"thrust_min":0,"thrust_max":215,"throttleable":true,"isp_vac":300,"isp_atm":280},
-			{"name":"LV-909 \"Terrier\" Liquid Fuel Engine","type":TYPES.LFO_ENGINE,"size":1,"cost":390,"mass":0.5,"thrust_min":0,"thrust_max":60,"throttleable":true,"isp_vac":345,"isp_atm":85,"gimbal":4},
-			{"name":"48-7S \"Spark\" Liquid Fuel Engine","type":TYPES.LFO_ENGINE,"size":0,"cost":200,"mass":0.1,"thrust_min":0,"thrust_max":18,"throttleable":true,"isp_vac":300,"isp_atm":270,"gimbal":3},
-			{"name":"LV-1 \"Ant\" Liquid Fuel Engine","type":TYPES.LFO_ENGINE,"size":0,"cost":110,"mass":0.02,"thrust_min":0,"thrust_max":2,"throttleable":true,"isp_vac":315,"isp_atm":85},
-			{"name":"Mk-55 \"Thud\" Liquid Fuel Engine","type":TYPES.LFO_ENGINE,"size":-1,"cost":820,"mass":0.9,"thrust_min":0,"thrust_max":120,"throttleable":true,"isp_vac":305,"isp_atm":275,"gimbal":8,"last":true,"radial":true},
-			{"name":"24-77 \"Twitch\" Liquid Fuel Engine","type":TYPES.LFO_ENGINE,"size":-1,"cost":400,"mass":0.09,"thrust_min":0,"thrust_max":16,"throttleable":true,"isp_vac":290,"isp_atm":250,"gimbal":8,"last":true,"radial":true},
-			{"name":"LV-1R \"Spider\" Liquid Fuel Engine","type":TYPES.LFO_ENGINE,"size":-1,"cost":120,"mass":0.02,"thrust_min":0,"thrust_max":2,"throttleable":true,"isp_vac":290,"isp_atm":260,"gimbal":8,"last":true,"radial":true},
+			{"name":"S3 KS-25x4 \"Mammoth\" Liquid Fuel Engine","type":TYPES.LFO_ENGINE,"size":3,"cost":39000,"mass":15,"thrust_min":0,"thrust_max":3746.031746,"throttleable":true,"isp_vac":315,"isp_atm":295,"thrust_atm":3746.031746,"thrust_vac":4000,"gimbal":2,"last":true},
+			{"name":"Kerbodyne KR-2L+ \"Rhino\" Liquid Fuel Engine","type":TYPES.LFO_ENGINE,"size":3,"cost":21000,"mass":8.5,"thrust_min":0,"thrust_max":1000,"throttleable":true,"isp_vac":340,"isp_atm":170,"thrust_atm":1000,"thrust_vac":2000,"gimbal":4},
+			{"name":"RE-M3 \"Mainsail\" Liquid Engine","type":TYPES.LFO_ENGINE,"size":2,"cost":13000,"mass":6,"thrust_min":0,"thrust_max":1379.032258,"throttleable":true,"isp_vac":310,"isp_atm":285,"thrust_atm":1379.032258,"thrust_vac":1500,"gimbal":2},
+			{"name":"RE-I5 \"Skipper\" Liquid Fuel Engine","type":TYPES.LFO_ENGINE,"size":2,"cost":5300,"mass":3,"thrust_min":0,"thrust_max":568.75,"throttleable":true,"isp_vac":320,"isp_atm":280,"thrust_atm":568.75,"thrust_vac":650,"gimbal":2},
+			{"name":"RE-L10 \"Poodle\" Liquid Fuel Engine","type":TYPES.LFO_ENGINE,"size":2,"cost":1300,"mass":1.75,"thrust_min":0,"thrust_max":53.42857143,"throttleable":true,"isp_vac":350,"isp_atm":85,"thrust_atm":53.42857143,"thrust_vac":220,"gimbal":4.5},
+			{"name":"CR-7 R.A.P.I.E.R. Engine","type":TYPES.LFO_ENGINE,"size":1,"cost":6000,"mass":2,"thrust_min":0,"thrust_max":162.3,"throttleable":true,"isp_vac":305,"isp_atm":275,"thrust_atm":162.3,"thrust_vac":180,"gimbal":3},
+			{"name":"LV-T45 \"Swivel\" Liquid Fuel Engine","type":TYPES.LFO_ENGINE,"size":1,"cost":1200,"mass":1.5,"thrust_min":0,"thrust_max":168.75,"throttleable":true,"isp_vac":320,"isp_atm":270,"thrust_atm":168.75,"thrust_vac":200,"gimbal":3},
+			{"name":"T-1 Toroidal \"Aerospike\" Liquid Fuel Engine","type":TYPES.LFO_ENGINE,"size":1,"cost":3850,"mass":1.5,"thrust_min":0,"thrust_max":153.788,"throttleable":true,"isp_vac":330,"isp_atm":290,"thrust_atm":153.788,"thrust_vac":175,"last":true},
+			{"name":"LV-T30 \"Reliant\" Liquid Fuel Engine","type":TYPES.LFO_ENGINE,"size":1,"cost":1100,"mass":1.25,"thrust_min":0,"thrust_max":200.6666667,"throttleable":true,"isp_vac":300,"isp_atm":280,"thrust_atm":200.6666667,"thrust_vac":215},
+			{"name":"LV-909 \"Terrier\" Liquid Fuel Engine","type":TYPES.LFO_ENGINE,"size":1,"cost":390,"mass":0.5,"thrust_min":0,"thrust_max":14.7826087,"throttleable":true,"isp_vac":345,"isp_atm":85,"thrust_atm":14.7826087,"thrust_vac":60,"gimbal":4},
+			{"name":"48-7S \"Spark\" Liquid Fuel Engine","type":TYPES.LFO_ENGINE,"size":0,"cost":200,"mass":0.1,"thrust_min":0,"thrust_max":16.2,"throttleable":true,"isp_vac":300,"isp_atm":270,"thrust_atm":16.2,"thrust_vac":18,"gimbal":3},
+			{"name":"LV-1 \"Ant\" Liquid Fuel Engine","type":TYPES.LFO_ENGINE,"size":0,"cost":110,"mass":0.02,"thrust_min":0,"thrust_max":0.54,"throttleable":true,"isp_vac":315,"isp_atm":85,"thrust_atm":0.54,"thrust_vac":2},
+			{"name":"Mk-55 \"Thud\" Liquid Fuel Engine","type":TYPES.LFO_ENGINE,"size":-1,"cost":820,"mass":0.9,"thrust_min":0,"thrust_max":108.1967213,"throttleable":true,"isp_vac":305,"isp_atm":275,"thrust_atm":108.1967213,"thrust_vac":120,"gimbal":8,"last":true,"radial":true},
+			{"name":"24-77 \"Twitch\" Liquid Fuel Engine","type":TYPES.LFO_ENGINE,"size":-1,"cost":400,"mass":0.09,"thrust_min":0,"thrust_max":13.793,"throttleable":true,"isp_vac":290,"isp_atm":250,"thrust_atm":13.793,"thrust_vac":16,"gimbal":8,"last":true,"radial":true},
+			{"name":"LV-1R \"Spider\" Liquid Fuel Engine","type":TYPES.LFO_ENGINE,"size":-1,"cost":120,"mass":0.02,"thrust_min":0,"thrust_max":1.793,"throttleable":true,"isp_vac":290,"isp_atm":260,"thrust_atm":1.793,"thrust_vac":2,"gimbal":8,"last":true,"radial":true},
 			
 			//LF/O Tanks
 			{"name":"Kerbodyne S3-14400 Tank","type":TYPES.LFO_TANK,"size":3,"cost":13000,"mass":9,"mass_fuel":72},
@@ -102,13 +102,12 @@ var PACKS = [
 			{"name":"Oscar-B Fuel Tank","type":TYPES.LFO_TANK,"size":0,"cost":70,"mass":0.025,"mass_fuel":0.2},
 			
 			//Boosters
-			{"name":"LFB KR-1x2 \"Twin-Boar\" Liquid Fuel Engine","type":TYPES.BOOSTER,"size":2,"cost":17000,"mass":10,"thrust_min":0,"thrust_max":2000,"throttleable":true,"isp_vac":300,"isp_atm":280,"gimbal":1.5,"mass_fuel":32,"last":true},
-			{"name":"S1 SRB-KD25k \"Kickback\" Solid Fuel Booster","type":TYPES.BOOSTER,"size":1,"cost":2700,"mass":4.5,"thrust_min":0,"thrust_max":671,"throttleable":false,"isp_vac":190,"isp_atm":170,"mass_fuel":18.75},
-			{"name":"BACC \"Thumper\" Solid Fuel Booster","type":TYPES.BOOSTER,"size":1,"cost":1050,"mass":1.5,"thrust_min":0,"thrust_max":300,"throttleable":false,"isp_vac":180,"isp_atm":165,"mass_fuel":4.5},
-			{"name":"Launch Escape System","type":TYPES.BOOSTER,"size":1,"cost":1000,"mass":0.9,"thrust_min":0,"thrust_max":750,"throttleable":false,"isp_vac":180,"isp_atm":160,"mass_fuel":0.225},
-			{"name":"RT-10 \"Hammer\" Solid Fuel Booster","type":TYPES.BOOSTER,"size":1,"cost":580,"mass":0.75,"thrust_min":0,"thrust_max":227,"throttleable":false,"isp_vac":162,"isp_atm":150,"mass_fuel":2.25},
-			{"name":"RT-5 \"Flea\" Solid Fuel Booster","type":TYPES.BOOSTER,"size":1,"cost":320,"mass":0.45,"thrust_min":0,"thrust_max":193,"throttleable":false,"isp_vac":150,"isp_atm":140,"mass_fuel":0.75},
-			{"name":"Sepratron I","type":TYPES.BOOSTER,"size":1,"cost":75,"mass":0.0125,"thrust_min":0,"thrust_max":23,"throttleable":false,"isp_vac":154,"isp_atm":118,"mass_fuel":0.06,"last":true,"radial":true},
+			{"name":"LFB KR-1x2 \"Twin-Boar\" Liquid Fuel Engine","type":TYPES.BOOSTER,"size":2,"cost":17000,"mass":10,"thrust_min":0,"thrust_max":1866.666667,"throttleable":true,"isp_vac":300,"isp_atm":280,"thrust_atm":1866.666667,"thrust_vac":2000,"gimbal":1.5,"mass_fuel":32,"last":true},
+			{"name":"S1 SRB-KD25k \"Kickback\" Solid Fuel Booster","type":TYPES.BOOSTER,"size":1,"cost":2700,"mass":4.5,"thrust_min":0,"thrust_max":600,"throttleable":false,"isp_vac":190,"isp_atm":170,"thrust_atm":600,"thrust_vac":671,"mass_fuel":18.75},
+			{"name":"BACC \"Thumper\" Solid Fuel Booster","type":TYPES.BOOSTER,"size":1,"cost":1050,"mass":1.5,"thrust_min":0,"thrust_max":275,"throttleable":false,"isp_vac":180,"isp_atm":165,"thrust_atm":275,"thrust_vac":300,"mass_fuel":4.5},
+			{"name":"RT-10 \"Hammer\" Solid Fuel Booster","type":TYPES.BOOSTER,"size":1,"cost":580,"mass":0.75,"thrust_min":0,"thrust_max":210,"throttleable":false,"isp_vac":162,"isp_atm":150,"thrust_atm":210,"thrust_vac":227,"mass_fuel":2.25},
+			{"name":"RT-5 \"Flea\" Solid Fuel Booster","type":TYPES.BOOSTER,"size":1,"cost":320,"mass":0.45,"thrust_min":0,"thrust_max":180,"throttleable":false,"isp_vac":150,"isp_atm":140,"thrust_atm":180,"thrust_vac":193,"mass_fuel":0.75},
+			{"name":"Sepratron I","type":TYPES.BOOSTER,"size":-1,"cost":75,"mass":0.0125,"thrust_min":0,"thrust_max":18,"throttleable":false,"isp_vac":154,"isp_atm":118,"thrust_atm":18,"thrust_vac":23,"mass_fuel":0.06,"last":true,"radial":true},
 			
 			//Decouplers
 			{"name":"TR-38-D","type":TYPES.DECOUPLER,"size":3,"cost":600,"mass":0.8,"ejection_force":100},
@@ -132,7 +131,7 @@ var PACKS = [
 	}
 ];
 
-var NO_ENGINE = {name:"", type:TYPES.LFO_ENGINE, size:0, cost:0, mass:0, thrust_min:0, thrust_max:0, isp_vac:0, isp_atm:0};
+var NO_ENGINE = {name:"", type:TYPES.LFO_ENGINE, size:0, cost:0, mass:0, thrust_min:0, thrust_max:0, isp_vac:0, isp_atm:0, thrust_atm:0, thrust_vac:0};
 var FUEL_DUCT = {name:"FTX-2 External Fuel Duct", type:TYPES.DUCT, size:"radial", cost:650, mass:0, radial:true};  //mass:0.05, but have no mass in flight
 
 var CONVERSION_FACTOR = 9.82;  //m/s^2
@@ -195,17 +194,17 @@ var KSP = {
 		//Engine fuel consumption at specified atmospheric pressure (t/s)
 		consumption : function (atm, engine) {
 			atm = Math.min(Math.max(atm, 0), 1);  //Atmospheric pressure's affect on engines maxes out at 1 (TODO: Confirm this)
-			return (engine.thrust_max / ((engine.isp_vac * CONVERSION_FACTOR) - ((engine.isp_vac - engine.isp_atm) * CONVERSION_FACTOR * (atm || 0)))) || 0;
+			return ((atm ? engine.thrust_atm : engine.thrust_vac) / ((engine.isp_vac * CONVERSION_FACTOR) - ((engine.isp_vac - engine.isp_atm) * CONVERSION_FACTOR * (atm || 0)))) || 0;
 		},
 		
 		//Engine fuel consumption in vacuum (t/s)
 		consumptionVac : function (engine) {
-			return (engine.thrust_max / (engine.isp_vac * CONVERSION_FACTOR)) || 0;
+			return (engine.thrust_vac / (engine.isp_vac * CONVERSION_FACTOR)) || 0;
 		},
 		
 		//Engine fuel consumption in atmosphere (t/s)
 		consumptionAtm : function (engine) {
-			return (engine.thrust_max / (engine.isp_atm * CONVERSION_FACTOR)) || 0;
+			return (engine.thrust_atm / (engine.isp_atm * CONVERSION_FACTOR)) || 0;
 		}
 	},
 	
@@ -235,10 +234,10 @@ var KSP = {
 		},
 		
 		//Thrust of active engines during this stage only (kN)
-		thrust : function (stage) {
+		thrust : function (stage, atm) {
 			return (
-				(stage.lfoEngines ? stage.lfoEngines.map(pluckNumber.bind(this, "thrust_max")).reduce(sum, 0) : 0) + 
-				(stage.boosters ? stage.boosters.map(pluckNumber.bind(this, "thrust_max")).reduce(sum, 0) : 0)
+				(stage.lfoEngines ? stage.lfoEngines.map(pluckNumber.bind(this, (atm ? "thrust_atm" : "thrust_vac"))).reduce(sum, 0) : 0) + 
+				(stage.boosters ? stage.boosters.map(pluckNumber.bind(this, (atm ? "thrust_atm" : "thrust_vac"))).reduce(sum, 0) : 0)
 			) * (stage.multiplier || 1);
 		},
 		
@@ -313,7 +312,7 @@ var KSP = {
 		
 		//Combined specific impulse at atmospheric pressure (m/s)
 		impulse : function (stage, atm) {
-			return (KSP.Stage.thrust(stage) / KSP.Stage.consumption(stage, atm)) || 0;
+			return (KSP.Stage.thrust(stage, atm) / KSP.Stage.consumption(stage, atm)) || 0;
 		},
 		
 		//Time required to use all fuel in this stage only at full thrust at specified atmospheric pressure (s)
@@ -336,9 +335,9 @@ var KSP = {
 		},
 		
 		//Thrust of all active engines during this stage (kN)
-		thrust : function (stage) {
-			return KSP.Stage.thrust(stage) +
-				(stage.parallel ? KSP.Stages.thrust(stage.next) : 0);
+		thrust : function (stage, atm) {
+			return KSP.Stage.thrust(stage, atm) +
+				(stage.parallel ? KSP.Stages.thrust(stage.next, atm) : 0);
 		},
 		
 		//Total mass of all stages (t)
@@ -390,8 +389,8 @@ var KSP = {
 		},
 		
 		//Thrust-to-Weight Ratio
-		twr : function (stage, planet) {
-			return (KSP.Stages.thrust(stage) / (KSP.Stages.massStart(stage) * planet.gravity)) || 0;
+		twr : function (stage, atm, planet) {
+			return (KSP.Stages.thrust(stage, atm) / (KSP.Stages.massStart(stage) * planet.gravity)) || 0;
 		},
 		
 		//Total stage fuel consumption at specified atmospheric pressure (t/s)
@@ -402,7 +401,7 @@ var KSP = {
 		
 		//Combined specific impulse at atmospheric pressure (m/s)
 		impulse : function (stage, atm) {
-			return (KSP.Stages.thrust(stage) / KSP.Stages.consumption(stage, atm)) || 0;
+			return (KSP.Stages.thrust(stage, atm) / KSP.Stages.consumption(stage, atm)) || 0;
 		},
 		
 		//Delta-V at atmospheric pressure of this stage only (m/s)
@@ -449,7 +448,7 @@ var KSP = {
 			if (stage.others && stage.others.length) strBuild.push("Other: " + KSP.Parts.humanize(stage.others));
 			if ((stage.lfoEngines && stage.lfoEngines.length) || (stage.boosters && stage.boosters.length)) {
 				if (typeof atm === "number") strBuild.push("Delta-V: " + Math.round(KSP.Stages.deltaVStage(stage, atm)));
-				if (planet) strBuild.push("TWR: " + KSP.Stages.twr(stage, planet).toFixed(2));
+				if (planet) strBuild.push("TWR: " + KSP.Stages.twr(stage, atm, planet).toFixed(2));
 			}
 			if (stage.parallel) {
 				if (stage.asparagus) strBuild.push("ASPARAGUS");
@@ -668,7 +667,7 @@ function findOptimalStage(args) {
 								}
 							}
 							
-							var twr = KSP.Stages.twr(stage, args.planet);
+							var twr = KSP.Stages.twr(stage, args.atm, args.planet);
 							var time = KSP.Stages.timeStage(stage, args.atm);
 							
 							if (twr < args.minTWR || time > args.maxTime) {
@@ -800,7 +799,7 @@ function findOptimalStage(args) {
 				}
 			}
 			
-			var twr = KSP.Stages.twr(stage, args.planet);
+			var twr = KSP.Stages.twr(stage, args.atm, args.planet);
 			var time = KSP.Stages.timeStage(stage, args.atm);
 			if (twr > args.maxTWR || time < args.minTime || time > args.maxTime) {
 				continue nextBooster;